Massive array of tables to be deleted in pheno schema.
Ultimately led to removal of all of the old pheno code, and some refactoring of the new code to make it use corect names.  A lot of code is removed and may eventually need to be replaced.  But basically all the old functionality that was there previously still remains, now need to complete pheno upload